# backend/docs/openapi.yaml
definitions:
...
  models.Partner:
    properties:
      address:
        description: Mailing address
        example: 5 Avenue Anatole France, Paris
        type: string
      created_at:
        description: Creation timestamp (UTC)
        format: date-time
        type: string
      description:
        description: Short description
        example: Sponsor
        type: string
      email:
        description: Contact email
        example: partners@acme.tld
        format: email
        type: string
      id:
        description: Unique partner identifier
        example: 1
        type: integer
      legal_status:
        description: Legal status
        example: SAS
        type: string
      name:
        description: Partner legal name
        example: ACME Corp
        type: string
      partnership_type:
        description: Type of partnership
        example: sponsor
        type: string
      phone:
        description: Contact phone
        example: +33 1 23 45 67 89
        type: string
    type: object
...
  requests.PartnerCreateRequest:
    properties:
      address:
        description: Mailing address
        example: 5 Avenue Anatole France, Paris
        type: string
      created_at:
        description: Partnership inception date/time
        format: date-time
        type: string
      description:
        description: Short description
        example: Long-term sponsor
        type: string
      email:
        description: Contact email
        example: partners@acme.tld
        format: email
        type: string
      legal_status:
        description: Legal status
        example: SAS
        type: string
      name:
        description: Partner legal name
        example: ACME Corp
        type: string
      partnership_type:
        description: Type of partnership
        example: sponsor
        type: string
      phone:
        description: Contact phone
        example: +33 1 23 45 67 89
        type: string
    type: object
  requests.PartnerUpdateRequest:
    properties:
      address:
        description: New address
        example: 10 Rue de Rivoli, Paris
        type: string
      created_at:
        description: New creation timestamp
        format: date-time
        type: string
      description:
        description: New description
        example: Updated description
        type: string
      email:
        description: New contact email
        example: partner@acme.tld
        format: email
        type: string
      legal_status:
        description: New legal status
        example: SASU
        type: string
      name:
        description: New legal name
        example: ACME Corp
        type: string
      partnership_type:
        description: New partnership type
        example: institutional
        type: string
      phone:
        description: New phone
        example: +33 6 12 34 56 78
        type: string
    type: object
...
  response.ErrorBody:
    properties:
      code:
        description: Machine-readable error code
        enum:
        - invalid_params
        - invalid_sort
        - invalid_payload
        - internal_error
        - not_found
        - email_taken
        - invalid_token
        - invalid_credentials
        - no_fields
        - method_not_allowed
        - unauthorized
        example: internal_error
        type: string
      details:
        description: Optional granular details (validation errors, etc.)
      message:
        description: Human-readable error message
        example: failed to retrieve resource
        type: string
    type: object
...
  response.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  response.PageMeta:
    properties:
      has_next:
        description: Whether there is a next page
        example: true
        type: boolean
      has_prev:
        description: Whether there is a previous page
        example: false
        type: boolean
      page:
        description: Current page number (1-based)
        example: 1
        minimum: 1
        type: integer
      per_page:
        description: Page size (items per page)
        example: 20
        minimum: 1
        type: integer
      total:
        description: Total number of items available
        example: 123
        minimum: 0
        type: integer
    type: object
  response.PartnerListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Partner'
        type: array
      pagination:
        $ref: '#/definitions/response.PageMeta'
    type: object
  response.PartnerObjectResponse:
    properties:
      data:
        $ref: '#/definitions/models.Partner'
    type: object
...
info:
  contact: {}
paths:
...
  /admin/partners:
    post:
      consumes:
      - application/json
      description: Creates a partner (admin required).
      parameters:
      - description: Partner
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.PartnerCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.PartnerObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Create partner
      tags:
      - Partners
  /admin/partners/{id}:
    delete:
      description: Deletes a partner (admin required).
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Delete partner
      tags:
      - Partners
    patch:
      consumes:
      - application/json
      description: Updates a partner (admin required).
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requests.PartnerUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PartnerObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      security:
      - BearerAuth: []
      summary: Update partner
      tags:
      - Partners
...
  /partners:
    get:
      description: Returns a paginated list of partners with sorting.
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: per_page
        type: integer
      - default: created_at
        description: Sort field
        enum:
        - id
        - name
        - email
        - legal_status
        - partnership_type
        - created_at
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PartnerListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: List partners
      tags:
      - Partners
  /partners/{id}:
    get:
      description: Retrieves a partner by ID.
      parameters:
      - description: Partner ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PartnerObjectResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorBody'
      summary: Get partner
      tags:
      - Partners
...
swagger: "2.0"